customer_support_agent/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pydantic>=2.0.0
typing-extensions>=4.0.0
